+++ source/features/pr-commit-lines-changed.tsx
@@ -6,7 +6,7 @@ import * as pageDetect from 'github-url-detection';
 import features from '../feature-manager.js';
 import api from '../github-helpers/api.js';
 import pluralize from '../helpers/pluralize.js';
-import GetCommitChanges from './get-commit-changes.gql';
+import GetCommitChanges from './pr-commit-lines-changed.gql';
 
 const commitChanges = new CachedFunction('commit-changes', {
 	async updater(commit: string): Promise<[additions: number, deletions: number]> {
